fix POST & PUT methods on invalid body

+++ routes/api/books/authors/delete.js
@@ -9,14 +9,16 @@ export const deleteBookAuthors = async (request, response) => {
 
         const author_ids = Array.isArray(request.body) ? request.body : [request.body];
 
+        if (author_ids.some(id => typeof id !== "number")) throw [400, errorCodes.RESOURCE_INVALID_REQUEST];
+
         await database.transaction(async trx => {
             for (const author_id of author_ids) {
                 let query = trx("book_authors").delete().where({ book_id, author_id });
 
-                logger.debug(`${request.originalUrl}: SQL: ${query.toString()}`)
+                logger.debug(`${request.method} ${request.originalUrl}: SQL: ${query.toString()}`)
         
                 await query.catch(error => {
-                    logger.error(`${request.originalUrl}: database error: ${query.toString()}: ${error}`);
+                    logger.error(`${request.method} ${request.originalUrl}: database error: ${query.toString()}: ${error}`);
                     throw [503, errorCodes.DATABASE_ERROR];
                 });
             }
@@ -28,7 +30,7 @@ export const deleteBookAuthors = async (request, response) => {
         });
     }
     catch ([status, errorCode]) {
-        logger.warn(`${request.originalUrl}: [${status}]: ${errorCodes.asMessage(errorCode)}`);
+        logger.warn(`${request.method} ${request.originalUrl}: [${status}]: ${errorCodes.asMessage(errorCode)}`);
 
         response.status(status);
         response.send({
