+++ routes/api/books/authors/put.js
@@ -7,7 +7,9 @@ export const putBookAuthors = async (request, response) => {
     try {
         const book_id = request.params.book_id;
 
-        const author_ids = request.body;
+        const author_ids = Array.isArray(request.body) ? request.body : [request.body];
+
+        if (author_ids.some(id => typeof id !== Number)) throw [400, errorCodes.RESOURCE_INVALID_REQUEST];
 
         await database.transaction(async trx => {
             await trx("book_authors").delete().where({ book_id });
@@ -15,14 +17,14 @@ export const putBookAuthors = async (request, response) => {
             for (const author_id of author_ids) {
                 let query = trx("book_authors").insert({ book_id, author_id });
 
-                logger.debug(`${request.originalUrl}: SQL: ${query.toString()}`)
+                logger.debug(`${request.method} ${request.originalUrl}: SQL: ${query.toString()}`)
         
                 const [inserted_id] = await query.catch(error => {
-                    logger.error(`${request.originalUrl}: database error: ${query.toString()}: ${error}`);
+                    logger.error(`${request.method} ${request.originalUrl}: database error: ${query.toString()}: ${error}`);
                     throw [503, errorCodes.DATABASE_ERROR];
                 });
         
-                logger.info(`${request.originalUrl}: inserted: ${inserted_id}`);
+                logger.info(`${request.method} ${request.originalUrl}: inserted: ${inserted_id}`);
             }
         });
 
@@ -32,7 +34,7 @@ export const putBookAuthors = async (request, response) => {
         });
     }
     catch ([status, errorCode]) {
-        logger.warn(`${request.originalUrl}: [${status}]: ${errorCodes.asMessage(errorCode)}`);
+        logger.warn(`${request.method} ${request.originalUrl}: [${status}]: ${errorCodes.asMessage(errorCode)}`);
 
         response.status(status);
         response.send({
